Former-commit-id: 5e8e0830892b7bc06413e710feb3a1e6c79022be

+++ installers/RPMs/yajsw-11.11/component.spec
@@ -70,7 +70,12 @@ cd /awips2/yajsw/lib/core/jna
 unzip -qu jna-3.4.1.jar com/sun/jna/linux-amd64/libjnidispatch.so
 
 %preun
-rm -fr /awips2/yajsw/lib/core/jna/com/
+
+#do not erase on an upgrade
+if [ "${1}" = "0" ]; then
+    rm -fr /awips2/yajsw/lib/core/jna/com/
+fi
+
 
 %clean
 rm -rf ${RPM_BUILD_ROOT}
